chore: moved enum on its own file
