+++ src/parser/github-comment-module.ts
@@ -2,10 +2,11 @@ import { Value } from "@sinclair/typebox/value";
 import Decimal from "decimal.js";
 import * as fs from "fs";
 import { stringify } from "yaml";
+import { CommentType } from "../configuration/comment-types";
 import configuration from "../configuration/config-reader";
 import { GithubCommentConfiguration, githubCommentConfigurationType } from "../configuration/github-comment-config";
 import { getOctokitInstance } from "../get-authentication-token";
-import { CommentType, IssueActivity } from "../issue-activity";
+import { IssueActivity } from "../issue-activity";
 import { parseGitHubUrl } from "../start";
 import { getPayoutConfigByNetworkId } from "../types/payout";
 import program from "./command-line";
